+++ include/slick/orderbook/detail/intrusive_list.hpp
@@ -46,6 +46,7 @@ public:
     using size_type = std::size_t;
 
     /// Forward iterator for intrusive list
+    /// We store a pointer to the list's tail to handle decrement from end()
     class Iterator {
     public:
         using iterator_category = std::bidirectional_iterator_tag;
@@ -54,8 +55,8 @@ public:
         using pointer = T*;
         using reference = T&;
 
-        constexpr Iterator() noexcept : current_(nullptr) {}
-        constexpr explicit Iterator(T* node) noexcept : current_(node) {}
+        constexpr Iterator() noexcept : current_(nullptr), tail_(nullptr) {}
+        constexpr Iterator(T* node, T* tail) noexcept : current_(node), tail_(tail) {}
 
         [[nodiscard]] constexpr reference operator*() const noexcept {
             return *current_;
@@ -77,7 +78,12 @@ public:
         }
 
         constexpr Iterator& operator--() noexcept {
-            current_ = current_->*PrevPtr;
+            // Handle decrement from end() (nullptr)
+            if (current_ == nullptr) {
+                current_ = tail_;
+            } else {
+                current_ = current_->*PrevPtr;
+            }
             return *this;
         }
 
@@ -95,8 +101,12 @@ public:
             return !(*this == other);
         }
 
+        // Allow ConstIterator to access private members
+        friend class ConstIterator;
+
     private:
         T* current_;
+        T* tail_;  // Needed to handle decrement from end() (nullptr)
     };
 
     /// Const iterator
@@ -108,9 +118,9 @@ public:
         using pointer = const T*;
         using reference = const T&;
 
-        constexpr ConstIterator() noexcept : current_(nullptr) {}
-        constexpr explicit ConstIterator(const T* node) noexcept : current_(node) {}
-        constexpr ConstIterator(Iterator it) noexcept : current_(it.operator->()) {}
+        constexpr ConstIterator() noexcept : current_(nullptr), tail_(nullptr) {}
+        constexpr ConstIterator(const T* node, const T* tail) noexcept : current_(node), tail_(tail) {}
+        constexpr ConstIterator(Iterator it) noexcept : current_(it.current_), tail_(it.tail_) {}
 
         [[nodiscard]] constexpr reference operator*() const noexcept {
             return *current_;
@@ -132,7 +142,12 @@ public:
         }
 
         constexpr ConstIterator& operator--() noexcept {
-            current_ = current_->*PrevPtr;
+            // Handle decrement from end() (nullptr)
+            if (current_ == nullptr) {
+                current_ = tail_;
+            } else {
+                current_ = current_->*PrevPtr;
+            }
             return *this;
         }
 
@@ -152,10 +167,139 @@ public:
 
     private:
         const T* current_;
+        const T* tail_;  // Needed to handle decrement from end() (nullptr)
+    };
+
+    /// Reverse iterator for intrusive list
+    /// Cannot use std::reverse_iterator because end() (nullptr) cannot be decremented
+    /// We store a pointer to the list's head to handle decrement from rend()
+    class ReverseIterator {
+    public:
+        using iterator_category = std::bidirectional_iterator_tag;
+        using value_type = T;
+        using difference_type = std::ptrdiff_t;
+        using pointer = T*;
+        using reference = T&;
+
+        constexpr ReverseIterator() noexcept : current_(nullptr), head_(nullptr) {}
+        constexpr ReverseIterator(T* node, T* head) noexcept : current_(node), head_(head) {}
+
+        [[nodiscard]] constexpr reference operator*() const noexcept {
+            return *current_;
+        }
+
+        [[nodiscard]] constexpr pointer operator->() const noexcept {
+            return current_;
+        }
+
+        constexpr ReverseIterator& operator++() noexcept {
+            current_ = current_->*PrevPtr;  // Move backward through list
+            return *this;
+        }
+
+        constexpr ReverseIterator operator++(int) noexcept {
+            ReverseIterator tmp = *this;
+            ++(*this);
+            return tmp;
+        }
+
+        constexpr ReverseIterator& operator--() noexcept {
+            // Handle decrement from rend() (nullptr)
+            if (current_ == nullptr) {
+                current_ = head_;
+            } else {
+                current_ = current_->*NextPtr;  // Move forward through list
+            }
+            return *this;
+        }
+
+        constexpr ReverseIterator operator--(int) noexcept {
+            ReverseIterator tmp = *this;
+            --(*this);
+            return tmp;
+        }
+
+        [[nodiscard]] constexpr bool operator==(const ReverseIterator& other) const noexcept {
+            return current_ == other.current_;
+        }
+
+        [[nodiscard]] constexpr bool operator!=(const ReverseIterator& other) const noexcept {
+            return !(*this == other);
+        }
+
+        // Allow ConstReverseIterator to access private members
+        friend class ConstReverseIterator;
+
+    private:
+        T* current_;
+        T* head_;  // Needed to handle decrement from rend() (nullptr)
+    };
+
+    /// Const reverse iterator
+    class ConstReverseIterator {
+    public:
+        using iterator_category = std::bidirectional_iterator_tag;
+        using value_type = T;
+        using difference_type = std::ptrdiff_t;
+        using pointer = const T*;
+        using reference = const T&;
+
+        constexpr ConstReverseIterator() noexcept : current_(nullptr), head_(nullptr) {}
+        constexpr ConstReverseIterator(const T* node, const T* head) noexcept : current_(node), head_(head) {}
+        constexpr ConstReverseIterator(ReverseIterator it) noexcept : current_(it.current_), head_(it.head_) {}
+
+        [[nodiscard]] constexpr reference operator*() const noexcept {
+            return *current_;
+        }
+
+        [[nodiscard]] constexpr pointer operator->() const noexcept {
+            return current_;
+        }
+
+        constexpr ConstReverseIterator& operator++() noexcept {
+            current_ = current_->*PrevPtr;
+            return *this;
+        }
+
+        constexpr ConstReverseIterator operator++(int) noexcept {
+            ConstReverseIterator tmp = *this;
+            ++(*this);
+            return tmp;
+        }
+
+        constexpr ConstReverseIterator& operator--() noexcept {
+            // Handle decrement from rend() (nullptr)
+            if (current_ == nullptr) {
+                current_ = head_;
+            } else {
+                current_ = current_->*NextPtr;
+            }
+            return *this;
+        }
+
+        constexpr ConstReverseIterator operator--(int) noexcept {
+            ConstReverseIterator tmp = *this;
+            --(*this);
+            return tmp;
+        }
+
+        [[nodiscard]] constexpr bool operator==(const ConstReverseIterator& other) const noexcept {
+            return current_ == other.current_;
+        }
+
+        [[nodiscard]] constexpr bool operator!=(const ConstReverseIterator& other) const noexcept {
+            return !(*this == other);
+        }
+
+    private:
+        const T* current_;
+        const T* head_;  // Needed to handle decrement from rend() (nullptr)
     };
 
     using iterator = Iterator;
     using const_iterator = ConstIterator;
+    using reverse_iterator = ReverseIterator;
+    using const_reverse_iterator = ConstReverseIterator;
 
     /// Constructor
     constexpr IntrusiveList() noexcept : head_(nullptr), tail_(nullptr), size_(0) {}
@@ -386,27 +530,52 @@ public:
 
     /// Iterators
     [[nodiscard]] constexpr iterator begin() noexcept {
-        return iterator(head_);
+        return iterator(head_, tail_);
     }
 
     [[nodiscard]] constexpr const_iterator begin() const noexcept {
-        return const_iterator(head_);
+        return const_iterator(head_, tail_);
     }
 
     [[nodiscard]] constexpr const_iterator cbegin() const noexcept {
-        return const_iterator(head_);
+        return const_iterator(head_, tail_);
     }
 
     [[nodiscard]] constexpr iterator end() noexcept {
-        return iterator(nullptr);
+        return iterator(nullptr, tail_);
     }
 
     [[nodiscard]] constexpr const_iterator end() const noexcept {
-        return const_iterator(nullptr);
+        return const_iterator(nullptr, tail_);
     }
 
     [[nodiscard]] constexpr const_iterator cend() const noexcept {
-        return const_iterator(nullptr);
+        return const_iterator(nullptr, tail_);
+    }
+
+    /// Reverse iterators
+    [[nodiscard]] constexpr reverse_iterator rbegin() noexcept {
+        return reverse_iterator(tail_, head_);
+    }
+
+    [[nodiscard]] constexpr const_reverse_iterator rbegin() const noexcept {
+        return const_reverse_iterator(tail_, head_);
+    }
+
+    [[nodiscard]] constexpr const_reverse_iterator crbegin() const noexcept {
+        return const_reverse_iterator(tail_, head_);
+    }
+
+    [[nodiscard]] constexpr reverse_iterator rend() noexcept {
+        return reverse_iterator(nullptr, head_);
+    }
+
+    [[nodiscard]] constexpr const_reverse_iterator rend() const noexcept {
+        return const_reverse_iterator(nullptr, head_);
+    }
+
+    [[nodiscard]] constexpr const_reverse_iterator crend() const noexcept {
+        return const_reverse_iterator(nullptr, head_);
     }
 
 private: